Implement switchable data types for many more fields

+++ TKMovesets2/Layouts/Editor/Structures/EditorCancels.cpp
@@ -57,7 +57,7 @@ void EditorCancels::OnUpdate(int listIdx, EditorInput* field)
 		if (m_editor->IsCommandGroupedCancelReference(commandField->buffer))
 		{
 			int groupId = atoi(moveIdField->buffer);
-			moveIdField->errored = groupId >= m_editor->GetStructureCount(EditorWindowType_GroupedCancel);
+			moveIdField->errored = groupId >= (int)m_editor->GetStructureCount(EditorWindowType_GroupedCancel);
 		}
 		else
 		{
@@ -94,7 +94,7 @@ void EditorCancels::BuildItemDetails(int listIdx)
 
 		std::string inputs;
 
-		if (inputSequenceId >= m_editor->GetStructureCount(EditorWindowType_InputSequence))
+		if (inputSequenceId >= (int)m_editor->GetStructureCount(EditorWindowType_InputSequence))
 		{
 			commandField->errored = true;
 			inputs = _("edition.cancel.invalid_sequence_id");
